HeartbeatError is sent blocking on shutdown
When the heartbeat error threshold is reached, and RMQ shuts itself
down, the final HeartbeatError is sent in a blocking manner.

We do this to ensure that if the client is listening for this error it
is guaranteed to eventually see the error. If this critical error is
dropped the client may be unaware that RMQ has stopped working and won't
shut itself down.