Ensure StopConsuming has completed for all calls
In tests we ensure that StopConsuming() successfully completes for all
tests.

The one small issue is inside TestMulti. Because the TestConsumer
instance waits for a call to Finish() in order to consume individual
messages in a controlled way, the consumer blocks waiting for calls to
Finish() after the main part of the test is complete.

We remedy this by adding a FinishAll() method to TestConsumer which
closes the `finsh` channel allowing the remaining messages to be
processed immediately.